requirements.txt
streamlit
nltk
torch
sentence-transformers
newspaper3k
ollama
requests
tqdm
ddgs
pandas